% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_MinimumSampleSize_fun.R
\name{CI_MinimumSampleSize_fun}
\alias{CI_MinimumSampleSize_fun}
\title{CI for minimum sample size}
\usage{
CI_MinimumSampleSize_fun(
  tr_set_size,
  prediction.ci,
  thr_acc,
  min_sam_size,
  fit_accuracy,
  w
)
}
\arguments{
\item{tr_set_size}{Training set size}

\item{prediction.ci}{Confidence Interval for the fitted curve, made up of three variables.}

\item{thr_acc}{Threshold of the chosen metric.}

\item{min_sam_size}{Estimated minimum sample size.}

\item{fit_accuracy}{Fitted object from nls.}

\item{w}{Width for CI estimation (may not be used)}
}
\value{
CI based on fitted upper and lower lines
}
\description{
Find the confidence interval for minimum sample size, given a fitted accuracy/Kappa curve
}
