% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acc_fun.R
\name{fit_acc_fun}
\alias{fit_acc_fun}
\title{Inverse function of the fitted accuracy or kappa}
\usage{
fit_acc_fun(a, b, c, x)
}
\arguments{
\item{a, b, c}{Coefficient from fitted formula}

\item{x}{Metric value (between 0 and 1)}
}
\value{
Corresponding sample size given the metric and parameters, solving
for the fitted formula.
}
\description{
Used to calculate the sample size, given the parameters and a metric value.
}
