% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_fun_parallel.R
\name{minimum_sample_knn}
\alias{minimum_sample_knn}
\title{kNN algorithm for minimum sample size estimation}
\usage{
minimum_sample_knn(X, Y, p_vec, param, thr_param, n.cores)
}
\arguments{
\item{X:}{variables to predict from}

\item{Y:}{predictor variables, i.e. the outcome}

\item{p_vec:}{a vector that gives the ratio of data to use in the training step}

\item{param:}{specify the metric to use. This can be "Accuracy" or "Kappa" for
classification problem.}

\item{thr_param:}{value of the metric for which to calculate the minimum sample size}

\item{n.cores:}{number of cores to use (for small datasets, recommended is 1)}
}
\value{
List with minimum sample size, corresponding CI, dataframe with sample size
and corresponding obtained metric, and fit parameters of the metric.
}
\description{
This algorithm determines the minimum sample size to use with the algorithm
knn, given a minimum value for the metric ("Accuracy" or "Kappa")
}
