% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_fun_v2_parallel.R
\name{minimum_sample_logistic_v2_parallel}
\alias{minimum_sample_logistic_v2_parallel}
\title{Logistic regression algorithm for minimum sample size estimation}
\usage{
minimum_sample_logistic_v2_parallel(X, Y, p_vec, param, thr_param, n.cores)
}
\description{
This algorithm determines the minimum sample size to use with the algorithm
logistic regresion, given a minimum value for the metric ("Accuracy" or "Kappa")
}
