% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_fun_parallel.R
\name{minimum_sample_naivebayes}
\alias{minimum_sample_naivebayes}
\title{Naive Bayes algorithm for minimum sample size estimation}
\usage{
minimum_sample_naivebayes(X, Y, p_vec, param, thr_param, n.cores)
}
\description{
This algorithm determines the minimum sample size to use with the algorithm
naive Bayes, given a minimum value for the metric ("Accuracy" or "Kappa")
}
