% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_fun_continuous.R
\name{minimum_sample_rf_continuous}
\alias{minimum_sample_rf_continuous}
\title{Random forest algorithm for minimum sample size estimation of regression}
\usage{
minimum_sample_rf_continuous(X, Y, p_vec, thr_R2, n.cores)
}
\description{
This algorithm determines the minimum sample size to use with the algorithm
random forest, given a minimum value for the metric ("R2")
}
