% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_fun_parallel.R
\name{minimum_sample_rf_parallel}
\alias{minimum_sample_rf_parallel}
\title{Random forest algorithm for minimum sample size estimation}
\usage{
minimum_sample_rf_parallel(X, Y, p_vec, param, thr_param, n.cores)
}
\description{
This algorithm determines the minimum sample size to use with the algorithm
random forest, given a minimum value for the metric ("Accuracy" or "Kappa")
}
