% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinSizeClassification.R
\name{MinSizeClassification}
\alias{MinSizeClassification}
\title{Algorithm for minimum sample size estimation in ML}
\usage{
MinSizeClassification(
  X,
  Y,
  algorithm,
  metric,
  thr_metric,
  p_vec = 1:99/100,
  n.cores = 1
)
}
\arguments{
\item{X}{Dataset of variable/s to use for prediction.}

\item{Y}{Vector with the predictor variable, i.e., single variable to classify.}

\item{algorithm}{Choose from "knn" (k Nearest Neighbors), "glm" (logistic regression), "nb" (Naive Bayes), or "rf" (Random Forest).
For binary classification any of the algorithms may be used. For multiple feature classification, only "knn" or "rf" may be used.}

\item{metric}{"Accuracy" or "Kappa". Classification metric to use to determine minimum sample size.}

\item{thr_metric}{Threshold on the metric to calculate the corresponding minimum sample size.}

\item{p_vec}{Vector of ratios to divide training data into. 
Code loops through the different ratios to get a sample size and calculate the corresponding metric. 
Default is 1:99/100}

\item{n.cores}{Number of logical CPU threads to use. Default is 1.}
}
\value{
List with minimum sample size, corresponding CI, dataframe with sample size, 
corresponding obtained metrics, and fit parameters of the metric.
}
\description{
This algorithm determines the minimum sample size to use with a specified algorithm,
given a minimum value for the metric ("Accuracy" or "Kappa").
It may be used for binary or multiple-feature classification.
}
