% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinSizeRFcontinuous.R
\name{MinSizeRFcontinuous}
\alias{MinSizeRFcontinuous}
\title{Random forest algorithm for minimum sample size estimation of regression}
\usage{
MinSizeRFcontinuous(X, Y, thr_R2, p_vec = 1:99/100, n.cores = 1)
}
\arguments{
\item{X}{Dataset of variable/s to use for prediction.}

\item{Y}{Vector with the predictor variable, i.e., single numeric variable to predict}

\item{thr_R2}{Threshold on the R2 metric to calculate the corresponding minimum sample size.}

\item{p_vec}{Vector of ratios to divide training data into. 
Code loops through the different ratios to get a sample size and calculate the corresponding metric. 
Default is 1:99/100}

\item{n.cores}{Number of logical CPU threads to use. Default is 1.}
}
\value{
List with minimum sample size, corresponding CI, dataframe with sample size, 
corresponding obtained metrics, and fit parameters of the metric.
}
\description{
This algorithm determines the minimum sample size to use with the algorithm
random forest, given a minimum value for the metric ("R2")
}
