% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acc_fun.R
\name{fit_acc_fun}
\alias{fit_acc_fun}
\title{Inverse function of the fitted accuracy or kappa}
\usage{
fit_acc_fun(a, b, c, x)
}
\description{
Used to calculate the corresponding size, given the parameters and an accuracy
or Kappa value.
}
